requirements.txt
Gunicorn
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.1
exceptiongroup==1.2.0
Flask==3.0.2
frozenlist==1.4.1
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
MouseInfo==0.1.3
multidict==6.0.5
Nuitka==2.0.1
numpy==1.24.4
openai==0.28.0
ordered-set==4.1.0
outcome==1.3.0.post0
pandas==2.0.3
pillow==10.2.0
playsound==1.3.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pyttsx3==2.90
pytweening==1.0.7
pytz==2024.1
pywin32==306
randfacts==0.21.0
requests==2.31.0
selenium==4.17.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SpeechRecognition==3.10.1
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
word2number==1.1
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zstandard==0.22.0